消除告警

+++ frameworks/native/pulseaudio/modules/abstract/module_loopback.c
@@ -557,7 +557,7 @@ static void MemblockqAdjust(struct userdata *u, int64_t latencyOffsetUsec, bool 
      * never underruns initially */
     pa_usec_t requestedSinkLatency = pa_sink_get_requested_latency_within_thread(u->sink_input->sink);
     if (requestedBufferLatency < (int64_t)requestedSinkLatency)
-        requestedBufferLatency = requestedSinkLatency;
+        requestedBufferLatency = (int64_t)requestedSinkLatency;
 
     size_t requestedMemblockqLength = pa_usec_to_bytes(requestedBufferLatency, &u->sink_input->sample_spec);
     size_t currentMemblockqLength = pa_memblockq_get_length(u->memblockq);
